backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
coverage==7.2.7
Django==3.2.25
django-cors-headers==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==18.13.0
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.2.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
six==1.17.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
zipp==3.15.0
